requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
mistralai==0.3.0
aiofiles==23.2.1
pydub==0.25.1
requests==2.31.0
python-dateutil==2.8.2
soundfile==0.12.1
